+++ packages/router/test/globalTestHook.ts
@@ -49,6 +49,8 @@ export const mochaHooks = {
       fulfill: stub().resolves(txReceiptMock),
       prepare: stub().resolves(txReceiptMock),
       removeLiquidity: stub().resolves(txReceiptMock),
+      getRouterBalance: stub().resolves(BigNumber.from("10001000000000000000000")),
+      sanitationCheck: stub().resolves(),
     };
 
     ctxMock = {