+++ packages/router/test/lib/operations/auction.spec.ts
@@ -36,7 +36,11 @@ describe("Auction Operation", () => {
   const { newAuction } = getOperations();
   describe("#newAuction", () => {
     beforeEach(() => {
-      getReceiverAmountStub = stub(PrepareHelperFns, "getReceiverAmount").resolves(MUTATED_AMOUNT);
+      getReceiverAmountStub = stub(PrepareHelperFns, "getReceiverAmount").resolves({
+        receivingAmount: MUTATED_AMOUNT,
+        routerFee: "10",
+        amountAfterSwapRate: MUTATED_AMOUNT,
+      });
 
       stub(PrepareHelperFns, "getReceiverExpiryBuffer").returns(MUTATED_BUFFER);
 
@@ -57,7 +61,11 @@ describe("Auction Operation", () => {
     });
 
     it("should error if not enough available liquidity for auction", async () => {
-      getReceiverAmountStub.returns("10002000000000000000000");
+      getReceiverAmountStub.returns({
+        receivingAmount: "10002000000000000000000",
+        routerFee: "10",
+        amountAfterSwapRate: MUTATED_AMOUNT,
+      });
       await expect(newAuction(auctionPayload, requestContext)).to.be.rejectedWith("Not enough liquidity");
     });
 
@@ -91,7 +99,11 @@ describe("Auction Operation", () => {
     });
 
     it("happy-1: should take a gas fee for fulfill transactions if sendingChain is fee chain", async () => {
-      getReceiverAmountStub.returns("100000000000000000000");
+      getReceiverAmountStub.returns({
+        receivingAmount: "100000000000000000000",
+        routerFee: "10",
+        amountAfterSwapRate: MUTATED_AMOUNT,
+      });
 
       // sendingChain = 1337, receivingChain = 1338
       getChainIdForGasFeeStub.returns([1337]);
@@ -127,14 +139,18 @@ describe("Auction Operation", () => {
 
     it("happy-2: should take a gas fee for prepare transactions if receivingChain is fee chain", async () => {
       AuctionHelperFns.AUCTION_REQUEST_MAP.clear();
-      getReceiverAmountStub.returns("100000000000000000000");
+      getReceiverAmountStub.returns({
+        receivingAmount: "100000000000000000000",
+        routerFee: "10",
+        amountAfterSwapRate: MUTATED_AMOUNT,
+      });
 
       // sendingChain = 1337, receivingChain = 1338
       getChainIdForGasFeeStub.returns([1338]);
 
       // it should take a gas fee for prepare transactions if receivingChain is fee chain.
       // amountReceived = amount.sub(prepareGasFee)
-      const expectedReceiverAmount = "99989500000000000000";
+      const expectedReceiverAmount = "99988800000000000000";
       const bid = await newAuction(auctionPayload, requestContext);
       expect(bid.bid).to.deep.eq({
         user: auctionPayload.user,
@@ -163,14 +179,18 @@ describe("Auction Operation", () => {
 
     it("happy-3: should take a gas fee for prepare and fulfill transactions if both sendingChain and receivingChain are fee chains", async () => {
       AuctionHelperFns.AUCTION_REQUEST_MAP.clear();
-      getReceiverAmountStub.returns("100000000000000000000");
+      getReceiverAmountStub.returns({
+        receivingAmount: "100000000000000000000",
+        routerFee: "10",
+        amountAfterSwapRate: MUTATED_AMOUNT,
+      });
 
       // sendingChain = 1337, receivingChain = 1338
       getChainIdForGasFeeStub.returns([1337, 1338]);
 
       // it should take a gas fee for prepare and fulfill transactions if both sendingChain and receivingChain are fee chains.
       // amountReceived = amount.sub(prepareGasFee).sub(fulfillGasFee)
-      const expectedReceiverAmount = "99976900000000000000";
+      const expectedReceiverAmount = "99976200000000000000";
       const bid = await newAuction(auctionPayload, requestContext);
       expect(bid.bid).to.deep.eq({
         user: auctionPayload.user,
@@ -199,7 +219,12 @@ describe("Auction Operation", () => {
 
     it("happy-4: shouldn't take a gas fee if both sendingChain and receivingChain aren't fee chains", async () => {
       AuctionHelperFns.AUCTION_REQUEST_MAP.clear();
-      getReceiverAmountStub.returns("100000000000000000000");
+
+      getReceiverAmountStub.returns({
+        receivingAmount: "100000000000000000000",
+        routerFee: "10",
+        amountAfterSwapRate: MUTATED_AMOUNT,
+      });
 
       // sendingChain = 1337, receivingChain = 1338
       getChainIdForGasFeeStub.returns([1]);