+++ packages/router/test/lib/operations/cancel.spec.ts
@@ -1,9 +1,17 @@
-import { expect, invariantDataMock, txReceiptMock, createLoggingContext, mkBytes32 } from "@connext/nxtp-utils";
+import {
+  expect,
+  invariantDataMock,
+  txReceiptMock,
+  createLoggingContext,
+  mkBytes32,
+  getNtpTimeSeconds,
+} from "@connext/nxtp-utils";
 
 import { cancelInputMock } from "../../utils";
-import { contractReaderMock, contractWriterMock } from "../../globalTestHook";
-import { cancel } from "../../../src/lib/operations/cancel";
+import { contractReaderMock, contractWriterMock, txServiceMock } from "../../globalTestHook";
+import { cancel, SENDER_PREPARE_BUFFER_TIME } from "../../../src/lib/operations/cancel";
 import { SinonStub } from "sinon";
+import { SenderTxTooNew } from "../../../src/lib/errors/cancel";
 
 const { requestContext } = createLoggingContext("TEST", undefined, mkBytes32("0xabc"));
 
@@ -20,6 +28,23 @@ describe("Cancel Sender Operation", () => {
       const _cancelInputMock = { ...cancelInputMock, side: "buggy" };
       await expect(cancel(invariantDataMock, _cancelInputMock, requestContext)).to.eventually.be.rejectedWith(
         "Params invalid",
+      );
+    });
+
+    it("should error if sender prepare tx is too recent", async () => {
+      const _cancelInputMock = { ...cancelInputMock };
+      const preparedTime = Math.floor(Date.now() / 1000);
+      (contractReaderMock.getTransactionForChain as SinonStub).resolves(undefined);
+      txServiceMock.getBlock.resolves({
+        timestamp: preparedTime,
+      } as any);
+      await expect(cancel(invariantDataMock, _cancelInputMock, requestContext)).to.eventually.be.rejectedWith(
+        SenderTxTooNew.getMessage(
+          invariantDataMock.transactionId,
+          invariantDataMock.sendingChainId,
+          preparedTime,
+          await getNtpTimeSeconds(),
+        ),
       );
     });
 
@@ -44,7 +69,11 @@ describe("Cancel Sender Operation", () => {
     });
 
     it("happy: should cancel for sender chain", async () => {
+      const time = Math.floor(Date.now() / 1000) - SENDER_PREPARE_BUFFER_TIME - 500;
       (contractReaderMock.getTransactionForChain as SinonStub).resolves(undefined);
+      txServiceMock.getBlock.resolves({
+        timestamp: time,
+      } as any);
       const receipt = await cancel(invariantDataMock, cancelInputMock, requestContext);
 
       expect(receipt).to.deep.eq(txReceiptMock);