+++ packages/test-ui/src/App.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable require-jsdoc */
-import { ChainData } from "@connext/nxtp-utils";
+import { ChainData, getChainData } from "@connext/nxtp-utils";
 import { Button, Col, Row, Tabs, Typography } from "antd";
-import { providers, Signer, utils } from "ethers";
+import { providers, Signer } from "ethers";
 import { ReactElement, useEffect, useState } from "react";
 
 import "./App.css";
@@ -26,7 +26,7 @@ Object.entries(chainConfig).forEach(([chainId, { provider, subgraph, transaction
 function App(): ReactElement | null {
   const [web3Provider, setProvider] = useState<providers.Web3Provider>();
   const [signer, setSigner] = useState<Signer>();
-  const [chainData, setChainData] = useState<ChainData[]>([]);
+  const [chainData, setChainData] = useState<Map<string, ChainData>>();
 
   const connectMetamask = async () => {
     const ethereum = (window as any).ethereum;
@@ -57,8 +57,8 @@ function App(): ReactElement | null {
 
   useEffect(() => {
     const init = async () => {
-      const json = await utils.fetchJson("https://raw.githubusercontent.com/connext/chaindata/main/crossChain.json");
-      setChainData(json);
+      const data = await getChainData();
+      setChainData(data);
     };
     init();
   }, []);