+++ packages/txservice/src/chainreader.ts
@@ -3,8 +3,8 @@ import { createLoggingContext, Logger, RequestContext } from "@connext/nxtp-util
 
 import { TransactionServiceConfig, validateTransactionServiceConfig, DEFAULT_CONFIG, ChainConfig } from "./config";
 import { ReadTransaction } from "./types";
-import { TransactionServiceFailure } from "./error";
 import { ChainRpcProvider } from "./provider";
+import { ConfigurationError, ProviderNotConfigured } from "./error";
 
 /**
  * @classdesc Performs onchain reads with embedded retries.
@@ -82,7 +82,8 @@ export class ChainReader {
     if (result.isErr()) {
       throw result.error;
     } else {
-      return result.value;
+      // bump gas price
+      return result.value.mul(120).div(100);
     }
   }
 
@@ -133,6 +134,24 @@ export class ChainReader {
   }
 
   /**
+   * Gets a block
+   *
+   * @param chainId - The ID of the chain for which this call is related.
+   * @returns block representing the specified
+   */
+  public async getBlock(
+    chainId: number,
+    blockHashOrBlockTag: providers.BlockTag | Promise<providers.BlockTag>,
+  ): Promise<providers.Block | undefined> {
+    const result = await this.getProvider(chainId).getBlock(blockHashOrBlockTag);
+    if (result.isErr()) {
+      throw result.error;
+    } else {
+      return result.value;
+    }
+  }
+
+  /**
    * Gets a trsanction receipt by hash
    *
    * @param chainId - The ID of the chain for which this call is related.
@@ -158,9 +177,7 @@ export class ChainReader {
   protected getProvider(chainId: number): ChainRpcProvider {
     // Ensure that a signer, provider, etc are present to execute on this chainId.
     if (!this.providers.has(chainId)) {
-      throw new TransactionServiceFailure(
-        `No provider was found for chain ${chainId}! Make sure this chain's providers are configured.`,
-      );
+      throw new ProviderNotConfigured(chainId.toString());
     }
     return this.providers.get(chainId)!;
   }
@@ -179,7 +196,9 @@ export class ChainReader {
       const chain: ChainConfig = chains[chainId];
       // Ensure at least one provider is configured.
       if (chain.providers.length === 0) {
-        const error = new TransactionServiceFailure(`Provider configurations not found for chainID: ${chainId}`);
+        const error = new ConfigurationError({
+          providers,
+        });
         this.logger.error("Failed to create transaction service", context, methodContext, error.toJson(), {
           chainId,
           providers,